# feat: optimize fastQuo by Knuth Algorithm D