--- a/.codecov.yaml
+++ b/.codecov.yaml
@@ -1,0 +1,3 @@
+coverage:
+  status:
+    project: false