--- a/testdata/fuzz/FuzzDivDec/d1ac6afed7c33047
+++ b/testdata/fuzz/FuzzDivDec/d1ac6afed7c33047
@@ -1,0 +1,9 @@
+go test fuzz v1
+bool(true)
+uint64(1234567890123456789)
+uint64(90)
+byte('\x00')
+bool(true)
+uint64(44)
+uint64(55)
+byte('b')